fix: scope navigation queue to provider context
